poprawiony navbar
